week7/skeleton-based/requirements.txt
appdirs==1.4.4
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
fonttools==4.51.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1713279365350/work
importlib_resources==6.4.0
iopath==0.1.10
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.7.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1701735466804/work
packaging==24.0
pandas==2.0.3
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1615243974661/work
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.0.1
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
torch==1.12.1
-e git+https://github.com/Jho-Yonsei/HD-GCN@672da2c3d84c9ea41dab124b7be8b3f82a83b6b2#egg=torchlight&subdirectory=torchlight
torchpack==0.2.2
torchvision==0.13.1
tqdm==4.66.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1708239446578/work
wandb==0.16.6
yacs==0.1.8
zipp==3.18.1
